add rotue and integrate the blade with route
