--- views/hello.blade.php
@@ -1,1 +1,0 @@
-@extends('layout.admin_dashboard')
